+++ apps/commerce-streamer/src/main/java/com/loopers/batch/application/required/MvProductRankMonthlyRepository.java
@@ -1,8 +1,8 @@
-package com.loopers.application.required;
+package com.loopers.batch.application.required;
 
 import java.util.List;
 
-import com.loopers.domain.ranking.MvProductRankMonthly;
+import com.loopers.batch.domain.MvProductRankMonthly;
 
 public interface MvProductRankMonthlyRepository {
     void saveAll(List<MvProductRankMonthly> rankings);
