+++ apps/commerce-streamer/src/main/java/com/loopers/batch/application/required/MvProductRankWeeklyRepository.java
@@ -1,8 +1,8 @@
-package com.loopers.application.required;
+package com.loopers.batch.application.required;
 
 import java.util.List;
 
-import com.loopers.domain.ranking.MvProductRankWeekly;
+import com.loopers.batch.domain.MvProductRankWeekly;
 
 public interface MvProductRankWeeklyRepository {
     void saveAll(List<MvProductRankWeekly> rankings);
