+++ apps/commerce-streamer/src/main/java/com/loopers/batch/infrastructure/ProductRankWriter.java
@@ -1,13 +1,13 @@
-package com.loopers.batch;
+package com.loopers.batch.infrastructure;
 
 import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import com.loopers.application.required.MvProductRankMonthlyRepository;
-import com.loopers.application.required.MvProductRankWeeklyRepository;
-import com.loopers.domain.ranking.MvProductRankMonthly;
-import com.loopers.domain.ranking.MvProductRankWeekly;
+import com.loopers.batch.application.required.MvProductRankMonthlyRepository;
+import com.loopers.batch.application.required.MvProductRankWeeklyRepository;
+import com.loopers.batch.domain.MvProductRankMonthly;
+import com.loopers.batch.domain.MvProductRankWeekly;
 
 import lombok.RequiredArgsConstructor;
